Fix: abort 해결/ 기능이 좀 이상 기능 아직 미해결
